breaking change: extract TOML plugin into a separate package
