+++ e2e/cases/import-toml/rsbuild.config.ts
@@ -1,0 +1,5 @@
+import { pluginToml } from '@rsbuild/plugin-toml';
+
+export default {
+  plugins: [pluginToml()],
+};
