+++ packages/compat/webpack/src/shared.ts
@@ -34,7 +34,6 @@ export const applyDefaultPlugins = (plugins: Plugins) =>
     import('./plugins/sass').then((m) => m.pluginSass()),
     import('./plugins/less').then((m) => m.pluginLess()),
     plugins.bundleAnalyzer(),
-    plugins.toml(),
     plugins.yaml(),
     plugins.splitChunks(),
     plugins.startUrl?.(),
