+++ packages/plugin-toml/CHANGELOG.md
@@ -1,0 +1,1 @@
+# @rsbuild/plugin-toml
