+++ packages/plugin-toml/compiled/toml-loader/index.d.ts
