+++ packages/plugin-toml/modern.config.ts
@@ -1,0 +1,7 @@
+import moduleTools from '@modern-js/module-tools';
+import { buildConfigWithMjs } from '../../scripts/modern.base.config';
+
+export default {
+  plugins: [moduleTools()],
+  buildConfig: buildConfigWithMjs,
+};
