+++ packages/plugin-toml/tests/tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "extends": "@rsbuild/tsconfig/base",
+  "compilerOptions": {
+    "baseUrl": "./",
+    "types": ["vitest/globals"],
+    "paths": {
+      "@src/*": ["../src/*"],
+      "@/*": ["../src/provider/*"],
+      "@rsbuild/test-helper": ["../../vitest-helper/*"]
+    }
+  },
+  "include": ["./**/*.ts"]
+}
